Add type support trait
