+++ dds_derive/src/derive/mod.rs
@@ -4,5 +4,5 @@ pub mod dds_serialize_data;
 pub mod dds_type_xml;
 pub mod dynamic_type;
 pub mod enum_support;
+pub mod type_support;
 pub mod xtypes;
-pub mod xtypes_type_object;
